Add PoC for authentication by cookie
The idea is to restrict access to clocks. for this, a cookie is stored
at the creation of a clock. This is basically a testcase, if the browser
stores the cookie from the redirect, or if an additional mechanism is
needed.

In practice, it is probably more sensible to set the cookie during the
first page visit and store its value in the clock if needed.